requirements.txt
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
chumpy==0.70
idna==3.10
numpy==1.23.1
omegaconf==2.3.0
pandas==2.0.3
pathlib==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
six==1.17.0
soupsieve==2.6
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
trimesh==4.6.13